+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
-## [1.1.0 - 2026-09-07]
+## [1.1.0 - 2026-09-10]
 ### Added
 - Dockerfile
 - Single Spa
@@ -180,6 +180,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Form field placeholder
 - Mandatory field marker
 - Description collapse rules
+- Favicon as backend environment variable
 
 ### Removed
 - Application Form